Import LaserSpeed.py from github.com/K40Nano/K40Nano
This file was imported by scorchworks in the upstream project.  He
notes this in passing in the changelog, with enough details to
determine the exact reference.

This file is the k40nano/LaserSpeed.py from the time after the commit
https://github.com/K40Nano/K40Nano/commit/3aef62608a3335984e18ee73292f5c67b96661bc

In the original repository, this file had test coverage, which were
not imported by scorchworks.